Fix for "fixed" gpg command
Due to a side-effect of a bug-fix in the gpg command, it is now
necessary to pass --batch to gpg commands that you wouldn't expect
to allocate a /dev/tty device.

In general though, passing --batch to gpg commands run within a script
is probably a good practice to follow.