Adds 2.1.0.1-rc as a v2 option.
A new version of stack has been made available, so a new
Dockerfile has been created to allow us to use it.

GPG verification of the stack download has been removed in this commit,
as there is currently no signature file available, and the SHA256
file is formatted incorrectly for automated tool usage.  Manual
verification of the SHA256 confirms that the download is correct.