Adds 2.7.3 stack image with both ARM64 and AMD64
This is our first multi-arch image. Most of the steps to create the
manifest were done manually using `docker manifest` commands. They're
not yet automated in a script for tagging. The build itself is
consolidated in the Dockerfiles though.

A binary copy of stack 2.7.3 for ARM64 is included because one isn't yet
distributed by commerialhaskell.

Lastly, the arm64 image could be leaner (it's like 1.8gigs). I skipped
doing that for now as I'm not super familiar with using different FROM /
COPY commands in a Dockerfile