v7 - Ensures entrypoint is a login shell
We need the entrypoint to be a login shell - which it isn't by default -
so that the bashrc is run and paths are setup correctly.

I thought this could be done without specifying a loginshell entry point
and using .profile or .bash_profile but I had no luck with that
approach. :(